rfctr: merge DataTable into CrunchCube
CrunchCube is the only subclass of DataTable. Move its methods into
CrunchCube and remove the mixins/data_table module.
