+++ src/cr/cube/crunch_cube.py
@@ -16,15 +16,15 @@ from scipy.stats.contingency import expected_freq
 
 from cr.cube import ITEM_DIMENSION_TYPES
 from cr.cube.cube_slice import CubeSlice
+from cr.cube.dimension import Dimension
 from cr.cube.measures.index import Index
 from cr.cube.measures.scale_means import ScaleMeans
-from cr.cube.mixins.data_table import DataTable
 from cr.cube.util import lazyproperty
 
 np.seterr(divide='ignore', invalid='ignore')
 
 
-class CrunchCube(DataTable):
+class CrunchCube(object):
     """The main API object for manipulating Crunch.io cube-responses.
 
     This object provides the main API methods for working with cubes. The
@@ -63,7 +63,7 @@ class CrunchCube(DataTable):
         try:
             if not isinstance(response, dict):
                 response = json.loads(response)
-            super(CrunchCube, self).__init__(response.get('value', response))
+            self._cube = response.get('value', response)
         except TypeError:
             # If an unexpected type is provided raise descriptive exception.
             if not isinstance(response, dict):
@@ -73,6 +73,37 @@ class CrunchCube(DataTable):
                 ).format(type(response)))
 
         self.slices = self.get_slices()
+
+    @lazyproperty
+    def all_dimensions(self):
+        """Gets the dimensions of the crunch cube.
+
+        This function is internal, and is not mean to be used by ouside users
+        of the CrunchCube class. The main reason for this is the internal
+        representation of the different variable types (namely the MR and the
+        CA). These types have two dimensions each, but in the case of MR, the
+        second dimensions shouldn't be visible to the user. This function
+        returns such dimensions as well, since they're necessary for the
+        correct implementation of the functionality for the MR type.
+        The version that is mentioned to be used by users is the
+        property 'dimensions'.
+        """
+        entries = self._cube['result']['dimensions']
+        return [
+            (
+                # Multiple Response and Categorical Array variables have
+                # two subsequent dimensions (elements and selections). For
+                # this reason it's necessary to pass in both of them in the
+                # Dimension class init method. This is needed in order to
+                # determine the correct type (CA or MR). We only skip the
+                # two-argument constructor for the last dimension in the list
+                # (where it's not possible to fetch the subsequent one).
+                Dimension(entry)
+                if i + 1 >= len(entries)
+                else Dimension(entry, entries[i + 1])
+            )
+            for (i, entry) in enumerate(entries)
+        ]
 
     def as_array(self, include_missing=False, weighted=True, adjusted=False,
                  include_transforms_for_dims=None, prune=False, margin=False):
@@ -126,6 +157,31 @@ class CrunchCube(DataTable):
     def col_direction_axis(self):
         return self.ndim - 2
 
+    def count(self, weighted=True):
+        """Get cube's count with automatic weighted/unweighted selection."""
+        if weighted and self.is_weighted:
+            return sum(
+                self._cube['result']['measures'].get('count', {}).get('data')
+            )
+        return self._cube['result']['n']
+
+    @lazyproperty
+    def counts(self):
+        unfiltered = self._cube['result'].get('unfiltered')
+        filtered = self._cube['result'].get('filtered')
+        return unfiltered, filtered
+
+    def data(self, weighted, margin=False):
+        """Get the data in non-flattened shape.
+
+        Converts the flattened shape (original response) into non-flattened
+        shape (count of elements per cube dimension). E.g. for a CAT x CAT
+        cube, with 2 categories in each dimension (variable), we end up with
+        a ndarray of shape (2, 2).
+        """
+        values = self.flat_values(weighted, margin)
+        return np.array(values).reshape(self._shape)
+
     @lazyproperty
     def description(self):
         """Return the description of the cube."""
@@ -145,6 +201,30 @@ class CrunchCube(DataTable):
             if i not in self.mr_selections_indices
         ]
 
+    @lazyproperty
+    def filter_annotation(self):
+        """Get cube's filter annotation."""
+        return self._cube.get('filter_names', [])
+
+    def flat_values(self, weighted, margin=False):
+        """Return list of measure values as found in cube response.
+
+        If *weighted* is True, weighted counts are returned if present in the
+        cube. Otherwise, unweighted counts are returned. If *margin* is True,
+        counts are returned even if mean values are present, which may be
+        preferred for example when calculating a margin.
+        """
+        values = self._cube['result']['counts']
+        if self.has_means and not margin:
+            mean = self._cube['result']['measures'].get('mean', {})
+            values = mean.get('data', values)
+        elif weighted and self.is_weighted:
+            count = self._cube['result']['measures'].get('count', {})
+            values = count.get('data', values)
+        values = [(val if not type(val) is dict else np.nan)
+                  for val in values]
+        return values
+
     def get_slices(self, ca_as_0th=False):
         """Return list of :class:`.CubeSlice` objects.
 
@@ -160,6 +240,14 @@ class CrunchCube(DataTable):
             CubeSlice(self, i, ca_as_0th)
             for i, _ in enumerate(self.labels()[0])
         ]
+
+    @lazyproperty
+    def has_means(self):
+        """Check if cube has means."""
+        measures = self._cube.get('result', {}).get('measures')
+        if not measures:
+            return False
+        return measures.get('mean', None) is not None
 
     @lazyproperty
     def has_mr(self):
@@ -209,6 +297,18 @@ class CrunchCube(DataTable):
         types = {d.type for d in self.dimensions}
         ca_types = {'categorical_array', 'categorical'}
         return self.ndim == 2 and types == ca_types
+
+    @lazyproperty
+    def is_weighted(self):
+        """Check if the cube dataset is weighted."""
+        weighted = self._cube.get('query', {}).get('weight', None) is not None
+        weighted = weighted or self._cube.get('weight_var', None) is not None
+        weighted = weighted or self._cube.get('weight_url', None) is not None
+        weighted = weighted or (
+            self._cube['result']['counts'] !=
+            self._cube['result']['measures'].get('count', {}).get('data')
+        )
+        return weighted
 
     def labels(self, include_missing=False, include_transforms_for_dims=False):
         """Gets labels for each cube's dimension.
@@ -341,6 +441,13 @@ class CrunchCube(DataTable):
         return den
 
     @lazyproperty
+    def missing(self):
+        """Get missing count of a cube."""
+        if self.has_means:
+            return self._cube['result']['measures']['mean']['n_missing']
+        return self._cube['result'].get('missing')
+
+    @lazyproperty
     def mr_dim_ind(self):
         indices = [
             i for i, dim in enumerate(self.dimensions)
@@ -349,6 +456,27 @@ class CrunchCube(DataTable):
         if indices:
             return indices[0] if len(indices) == 1 else tuple(indices)
         return None
+
+    @lazyproperty
+    def mr_selections_indices(self):
+        """Gets indices of each 'selection' dim, for corresponding MR dim.
+
+        Multiple Response (MR) and Categorical Array (CA) variables are
+        represented by two dimensions each. These dimensions can be thought of
+        as 'elements' and 'selections'. This function returns the indices of
+        the 'selections' dimension for each MR variable.
+        """
+        mr_dimensions_indices = [
+            i for (i, dim) in enumerate(self.all_dimensions)
+            if (i + 1 < len(self.all_dimensions) and
+                dim.type == 'multiple_response')
+        ]
+
+        # For each MR and CA dimension, the 'selections' dimension
+        # follows right after it (in the originating cube).
+        # Here we increase the MR index by 1, which gives us
+        # the index of the corresponding 'selections' dimension.
+        return [i + 1 for i in mr_dimensions_indices]
 
     @lazyproperty
     def name(self):
@@ -1157,6 +1285,10 @@ class CrunchCube(DataTable):
             axis=axis, weighted=False,
             include_transforms_for_dims=hs_dims,
         )
+
+    @lazyproperty
+    def _shape(self):
+        return tuple([dim.shape for dim in self.all_dimensions])
 
     def _transform(self, res, include_transforms_for_dims,
                    inflate=False, fix=True):
